tendermint-rpc: Rearrange adaptor code
