+++ packages/tendermint-rpc/src/tendermint33/adaptor/index.ts
@@ -1,10 +1,12 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import { Adaptor } from "../../adaptor";
-import { hashBlock, hashTx } from "../../hasher";
+import { hashBlock, hashTx } from "../hasher";
 import { Params } from "./requests";
 import { Responses } from "./responses";
+import { Adaptor } from "./types";
 
-export const v0_33: Adaptor = {
+export { Decoder, Encoder, Params, Responses } from "./types";
+
+export const adaptor33: Adaptor = {
   params: Params,
   responses: Responses,
   hashTx: hashTx,
