--- packages/tendermint-rpc/src/tendermint33/adaptors/index.ts
@@ -1,2 +1,0 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-export { v0_33 as adaptor33 } from "./v0-33";
